semiconductor-search/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
openai>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.12
